move controllers to util directory
