+++ backend/adumbra/webserver/util/images.py
